+++ ext/Shared/Languages/Languages.lua
@@ -1,0 +1,5 @@
+Languages = {
+	"en_EN",
+	"de_DE",
+	"cn_CN"
+}